Criado mais 2 lixeiras dentro de prédios e terminado o dto para retornar as lixeiras foras de predio e contagem de lixeiras dentro de prédios

+++ src/main/java/pucrs/ages/garbus/dtos/BuildingsReduceDTO.java
@@ -14,6 +14,6 @@ public class BuildingsReduceDTO {
 
     private String name;
 
-    private int trashesCount;
+    private Long trashesCount;
 
 }
